+++ pages/portfolio/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 
 // Import components
-import PortfolioPage from "../../components/v1/PortfolioPage/PortfolioPageV2";
+import PortfolioPage from "../../modules/portofolio/PortofolioContainer";
 
 const PorfolioPage: NextPage = () => {
     return <PortfolioPage />;
